style: id인풋 스타일 변경
